KMNChatBot/requirements.txt
discord.py
newsapi-python
numpy
pandas
matplotlib
yfinance
pandas-datareader
requests
bs4
datetime